+++ packages/mui-components/prettier.config.js
@@ -1,1 +1,1 @@
-module.exports = require('@batoanng/prettier-config')
+module.exports = require('@batoanng/prettier-config');