+++ packages/mui-components/src/components/atoms/index.ts
@@ -2,3 +2,5 @@ export * from './FormErrorText';
 export * from './Button';
 export * from './SROnly';
 export * from './Loader';
+export * from './TextInput';
+export * from './HelpMessage';