+++ packages/mui-components/src/components/molecules/IconModal/index.ts
@@ -1,0 +1,3 @@
+export * from './IconModal';
+export * from './ModalIcon';
+export * from './CloseButton';