+++ packages/mui-components/src/components/molecules/index.ts
@@ -2,3 +2,5 @@ export * from './Notification';
 export * from './Error';
 export * from './IdleTimer';
 export * from './Modal';
+export * from './SlideInModal';
+export * from './IconModal';