+++ packages/mui-components/src/functions/index.ts
@@ -1,0 +1,2 @@
+export * from './replaceIntlDialingPrefixAu';
+export * from './getMaxWidth';