+++ packages/mui-components/src/index.ts
@@ -1,6 +1,9 @@
 /// <reference types="vite-plugin-svgr/client" />
 
 export * from './components';
+export * from './form';
 export * from './theme';
+export * from './functions';
 export * from './hooks';
 export * from './test-utils';
+export * from './types';